rename controllers and update submodule
