+++ src/models/Camera.js
@@ -1,5 +1,5 @@
 const dao = require('../dao/camera')
-const { getNodeIp } = require('../controllers/camera')
+const { getNodeIp } = require('../controllers/CameraController')
 const requestToNode = require('../node_client/NodeClient')
 const CameraConfigurations = require('./CameraConfigurations')
